#Add - Writes Enum as Text when Alphanumeric Picture is Provided
When setting an alphanumeric picture to X(n), then automatically its gonna
translate to Enum Name, not the integer.
Also, two test cases were provided in EdiTextWriterTests.cs.
This Implementation need to be analyzed because of changes in the ConvertUtils.cs, maybe we need to change de approach to some code
that Add Function not change the function.
This can be addressed to Issue #183